feat: migrate mobile suits admin page to /admin namespace
- Create Admin::MobileSuitsController (extends BaseController)
- Move views from app/views/mobile_suits/ to app/views/admin/mobile_suits/
- Update routes from /mobile_suits to /admin/mobile_suits
- Update all menu links to use admin_mobile_suits_path
