+++ app/views/admin/mobile_suits/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: mobile_suit, class: "space-y-6") do |form| %>
+<%= form_with(model: [:admin, mobile_suit], class: "space-y-6") do |form| %>
   <% if mobile_suit.errors.any? %>
     <div class="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded">
       <h2 class="font-bold text-lg mb-2"><%= pluralize(mobile_suit.errors.count, "個のエラー") %>があります:</h2>
@@ -26,7 +26,7 @@
   </div>
 
   <div class="flex justify-end space-x-3">
-    <%= link_to "キャンセル", mobile_suits_path, class: "px-4 py-2 bg-gray-600 text-white font-semibold rounded-md hover:bg-gray-700" %>
+    <%= link_to "キャンセル", admin_mobile_suits_path, class: "px-4 py-2 bg-gray-600 text-white font-semibold rounded-md hover:bg-gray-700" %>
     <%= form.submit mobile_suit.new_record? ? "登録" : "更新", class: "px-4 py-2 bg-indigo-600 text-white font-semibold rounded-md hover:bg-indigo-700" %>
   </div>
 <% end %>
