+++ app/views/admin/mobile_suits/new.html.erb
