+++ app/views/layouts/application.html.erb
@@ -61,7 +61,7 @@
                 </div>
               </div>
 
-              <%= link_to "機体マスタ", mobile_suits_path, class: "sidebar-nav-item #{current_page?(mobile_suits_path) ? 'active' : ''}" %>
+              <%= link_to "機体マスタ", admin_mobile_suits_path, class: "sidebar-nav-item #{current_page?(admin_mobile_suits_path) ? 'active' : ''}" %>
               <%= link_to "ユーザー管理", admin_users_path, class: "sidebar-nav-item #{current_page?(admin_users_path) ? 'active' : ''}" %>
             <% end %>
           </nav>
@@ -132,7 +132,7 @@
                 <%= link_to "対戦履歴", matches_path, class: "inline-flex items-center px-1 pt-1 text-sm font-medium text-gray-900 border-b-2 #{current_page?(matches_path) ? 'border-indigo-500' : 'border-transparent hover:border-gray-300'}" %>
                 <%= link_to "統計", statistics_path, class: "inline-flex items-center px-1 pt-1 text-sm font-medium text-gray-900 border-b-2 #{current_page?(statistics_path) ? 'border-indigo-500' : 'border-transparent hover:border-gray-300'}" %>
                 <% if current_user.is_admin? %>
-                  <%= link_to "機体マスタ", mobile_suits_path, class: "inline-flex items-center px-1 pt-1 text-sm font-medium text-gray-900 border-b-2 #{current_page?(mobile_suits_path) ? 'border-indigo-500' : 'border-transparent hover:border-gray-300'}" %>
+                  <%= link_to "機体マスタ", admin_mobile_suits_path, class: "inline-flex items-center px-1 pt-1 text-sm font-medium text-gray-900 border-b-2 #{current_page?(admin_mobile_suits_path) ? 'border-indigo-500' : 'border-transparent hover:border-gray-300'}" %>
                   <%= link_to "ユーザー管理", admin_users_path, class: "inline-flex items-center px-1 pt-1 text-sm font-medium text-gray-900 border-b-2 #{current_page?(admin_users_path) ? 'border-indigo-500' : 'border-transparent hover:border-gray-300'}" %>
                 <% end %>
               </div>
@@ -304,7 +304,7 @@
             <% if current_user.is_admin? %>
               <div class="border-t border-indigo-400 mt-3 pt-3">
                 <div class="px-4 py-2 text-xs font-semibold text-indigo-200 uppercase tracking-wider">管理者メニュー</div>
-                <%= link_to "機体マスタ", mobile_suits_path, class: "mobile-menu-item #{current_page?(mobile_suits_path) ? 'active' : ''}" %>
+                <%= link_to "機体マスタ", admin_mobile_suits_path, class: "mobile-menu-item #{current_page?(admin_mobile_suits_path) ? 'active' : ''}" %>
                 <%= link_to "ユーザー管理", admin_users_path, class: "mobile-menu-item #{current_page?(admin_users_path) ? 'active' : ''}" %>
               </div>
             <% end %>
