+++ config/routes.rb
@@ -37,9 +37,6 @@ Rails.application.routes.draw do
   # Statistics
   get "statistics", to: "statistics#index", as: :statistics
 
-  # Mobile Suits
-  resources :mobile_suits
-
   # Events
   resources :events do
     resources :matches, only: [:new, :create]
@@ -77,6 +74,7 @@ Rails.application.routes.draw do
         post :clear_view
       end
     end
+    resources :mobile_suits, except: [:show]
     resources :imports, only: [:new, :create] do
       collection do
         get 'mobile_suits/new', to: 'imports#new_mobile_suits', as: 'new_mobile_suits'
